requirements.txt
flask==3.1.0
flask-socketio==5.5.1
paho-mqtt==2.1.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2
gunicorn==21.2.0
eventlet==0.37.0
